Uniformly rollout discrete_extent throughout the code base
In the early days of the course, we passed width and height as separate parameters. Relatively recently, we introduced a discrete_extent type to wrap width and height. However, it was employed only for new code, leading to an insidious lack of harmonization across the code base. Here we fix this by deploying discrete_extent wherever we can.

An additional benefit beyond harmonization and simplification is that we are guided to use std::uint32_t more generally. By then computing image sizes using std::uint64_t we completely avoid the potential overflow issues we had in the past.

The excellent test coverage we have was instrumental in refactoring correctly. It picked up very subtle mistakes I made:

1. Exceptions not being thrown when they should have been due to overflow indicated that I had neglected to change std::size_t --> std::unit32_t in the signature of the multiplication_overflows.

2. By recording the text from exception messages, I caught that I had incorrectly updated impl::image_spec::validate.

+++ Tests/Core/AssetManagement/AlignmentTestingUtilities.hpp
@@ -23,7 +23,7 @@ namespace sequoia::testing
         }
         
         template<test_mode Mode>
-        static void test(equivalence_check_t, test_logger<Mode>& logger, const avocet::alignment& actual, const std::size_t prediction)
+        static void test(equivalence_check_t, test_logger<Mode>& logger, const avocet::alignment& actual, const std::uint32_t prediction)
         {
             check(equality, "Raw value", logger, actual.raw_value(), prediction);
         }
